x/jklmining/types/messages_miner_claims_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/jackal-dao/canine/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateMinerClaims_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateMinerClaims
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateMinerClaims{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateMinerClaims{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateMinerClaims_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateMinerClaims
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateMinerClaims{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateMinerClaims{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteMinerClaims_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteMinerClaims
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteMinerClaims{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteMinerClaims{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
